refactor: moved getPodsList to api
